Update Easy Enrol button to fix capability
The "enrol/category:config" is not necessary to use easy enrol integration and if in the Moodle, for any reason, we remove mentioned capability from users they will not be able to see the "Key" button on top of course top bar. Instead, if we change it to "enrol/easy:config" it will be work fine because it is a required capability to grant to any user who is eligible to manage easy enrol codes.